refactor: move pyx files in a mixmod folder
