+++ Pymixmod/mixmod/strategy.pxi
